karstd: don't truncate the scratch-dir hash, avoid the clippy lint
The previous commit's "as u32" truncation of the DefaultHasher output
tripped clippy::cast_possible_truncation under CI's actual invocation
(cargo clippy --workspace --all-targets --all-features -D warnings) --
a stricter flag set than the plain "cargo clippy -p karstd --lib" this
was checked against locally the first time, which stayed quiet. Using
the full u64 in hex sidesteps the cast entirely; it's still only 16
hex characters, well within the SUN_LEN budget the previous commit
already worked out (/tmp base, worst-case tag length included).

+++ bins/karstd/src/scratch.rs
@@ -75,7 +75,7 @@ impl Scratch {
         } else {
             std::env::temp_dir()
         };
-        let dir = base.join(format!("krst-{:x}-{tag}", hasher.finish() as u32));
+        let dir = base.join(format!("krst-{:x}-{tag}", hasher.finish()));
         let _ = std::fs::remove_dir_all(&dir);
         std::fs::create_dir_all(&dir).expect("create scratch directory");
         Self(dir)
